1.5.1 - Utilize OOP Javascript
-> Inheritance not working
I've plan to have 3 classes for this website :
1. FetchData()
2.FilmCarousel()
3.FilmCards()

I want to export FetchData() to other class for those to inheritance fetch class. However. the data doesn't pass on to FilmCarousel() and FilmCards() class. Therefore, it need to have fetch method for each class ( FilmCarousel(), FilmCards() )

-> search() not working in OOP

search method in FilmCards() class is not working. it said

" Uncaught ReferenceError: search is not defined
    at HTMLInputElement.onkeyup (films.html:87:64) "

Therefore, I put search function into top of html file in order for it to work

**Notes**

- Modal still not working
- animation for film card still need to be add for better looking